fix:Modify the test code name
