+++ client/purchase.go
@@ -26,7 +26,7 @@ pbk:wallet's public key
 func ObtainFromFaucet(pbk string) error {
 	pubkey, err := tools.DecodeToPub(pbk, tools.ChainCessTestPrefix)
 	if err != nil {
-		fmt.Printf("[Error]The wallet address you entered is incorrect, please re-enter\n")
+		fmt.Printf("[Error]The wallet address you entered is incorrect, please re-enter:%v\n", err.Error())
 		return err
 	}
 	pbk = fmt.Sprintf("%#x", pubkey)
