+++ install_ctl/cess_client.yaml
@@ -5,7 +5,6 @@ chainData:
   cessRpcAddr : ""
   faucetAddress : ""
   idAccountPhraseOrSeed : ""
-  accountPublicKey : ""
   walletAddress : ""
 
 pathInfo:
